The new content can be used but must not - this depends on the used TestStateStrategy.

Currently enabled: 
* when a word is added to the text state, a noun mapping and a phrase mapping is created
* noun mappings are stored in a set, compared via their reference and phrases
* phrase mappings are also stored in a set, compared via their phrases
* noun mappings and phrase mappings are not directly connected - the relation between them can only be obtained with the text state - this reduces effort for consistency control
* the text state merges no longer noun mappings - the text stage should be used to identify the relevant words. Noun mappings should be merged by later agents that run when it can be ensured that the whole text was read by all noun mapping creating agents. Whether this is at the end of the text stage or at the beginning of the recommendation stage has to be discussed.
* the mapping combiner agent (currently at the end of the text stage) merges noun mappings that are similar/ textually equal and have similar/ the same phrase mapping
* noun mappings are only created by the text state

New Extractors & Agents: 
* At the end of the text extraction stage - there is a new agent called Mapping Combiner. The agent combines several noun mappings if they and their phrases are similar.

Defintions and invariants: 
* every noun mapping has exactly one phrase mapping
* a phrase mapping has at least one noun mapping
* every word that was added to the text state is converted/ added to a noun mapping that can be obtained with getNounMappingByWord(word). For its phrase mapping this works analogue.


Future work: 
* discuss the position/ stage of merging agents
* are coreferences still needed? 
* refactor configurations for the new agents

+++ tests/src/test/java/edu/kit/kastel/mcse/ardoco/core/tests/integration/statehelper/ChangedStatesTest.java
@@ -1,3 +1,4 @@
+/* Licensed under MIT 2022. */
 package edu.kit.kastel.mcse.ardoco.core.tests.integration.statehelper;
 
 import java.io.File;
@@ -5,15 +6,10 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.*;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.slf4j.Logger;
@@ -23,62 +19,39 @@ import edu.kit.kastel.mcse.ardoco.core.api.output.ArDoCoResult;
 import edu.kit.kastel.mcse.ardoco.core.pipeline.ArDoCo;
 import edu.kit.kastel.mcse.ardoco.core.pipeline.ArchitectureModelType;
 import edu.kit.kastel.mcse.ardoco.core.tests.eval.Project;
+import edu.kit.kastel.mcse.ardoco.core.tests.integration.statehelper.files.ConnectionStateFile;
 import edu.kit.kastel.mcse.ardoco.core.tests.integration.statehelper.files.RecommendationStateFile;
 import edu.kit.kastel.mcse.ardoco.core.tests.integration.statehelper.files.TextStateFile;
-import edu.kit.kastel.mcse.ardoco.core.tests.integration.tlrhelper.TLProjectEvalResult;
 
+@Disabled("Disabled since it compares two testouts")
 public class ChangedStatesTest {
-    private static final boolean OVERWRITE_PREVIOUS = true;
+    private static final boolean OVERWRITE_PREVIOUS = false;
     private static final Logger logger = LoggerFactory.getLogger(ChangedStatesTest.class);
 
     private static final String OUTPUT = "src/test/resources/testout";
     private static final Path OUTPUT_PATH = Path.of(OUTPUT);
-    private static final String ADDITIONAL_CONFIG = null;
-    private static final List<TLProjectEvalResult> RESULTS = new ArrayList<>();
     private static final Map<Project, ArDoCoResult> DATA_MAP = new HashMap<>();
-    private static final boolean detailedDebug = true;
     private static final String LOGGING_ARDOCO_CORE = "org.slf4j.simpleLogger.log.edu.kit.kastel.mcse.ardoco.core";
-    private final File inputCodeModel = null;
-    private File additionalConfigs = null;
-    private final File outputDir = new File(OUTPUT);
 
     @BeforeAll
     public static void beforeAll() {
         System.setProperty(LOGGING_ARDOCO_CORE, "info");
     }
 
-    @AfterEach
-    void afterEach() {
-        if (ADDITIONAL_CONFIG != null) {
-            var config = new File(ADDITIONAL_CONFIG);
-            config.delete();
-        }
-        if (additionalConfigs != null) {
-            additionalConfigs = null;
-        }
-    }
+    @BeforeAll
+    static void evaluateStates() throws IOException {
+        for (Project project : Project.values()) {
+            File inputModel = project.getModelFile();
+            File inputText = project.getTextFile();
+            String name = project.name().toLowerCase();
+            ArDoCo arDoCo = ArDoCo.getInstance(name);
 
-    // NOTE: if you only want to test a specific project, you can simply set up the
-    // EnumSource. For more details, see
-    // https://www.baeldung.com/parameterized-tests-junit-5#3-enum
-    // Example: add ", names = { "BIGBLUEBUTTON" }" to EnumSource
-    // However, make sure to revert this before you commit and push!
-    @DisplayName("Evaluate Changed States (Text-based)")
-    @ParameterizedTest(name = "Evaluating {0} (Changed States)")
-    @EnumSource(value = Project.class)
-    void evaluateStates(Project project) throws IOException {
-        File inputModel = project.getModelFile();
-        File inputText = project.getTextFile();
-        String name = project.name().toLowerCase();
-        ArDoCo arDoCo = ArDoCo.getInstance(name);
-
-        var arDoCoResult = DATA_MAP.get(project);
-        if (arDoCoResult == null) {
-            arDoCoResult = arDoCo.runAndSave(name, inputText, inputModel, ArchitectureModelType.PCM, inputCodeModel, additionalConfigs, outputDir);
-            DATA_MAP.put(project, arDoCoResult);
+            var arDoCoResult = DATA_MAP.get(project);
+            if (arDoCoResult == null) {
+                arDoCoResult = arDoCo.runAndSave(name, inputText, inputModel, ArchitectureModelType.PCM, null, null, new File(OUTPUT));
+                DATA_MAP.put(project, arDoCoResult);
+            }
         }
-
-        writeDetailedOutput(project, arDoCoResult);
     }
 
     private static boolean writeTextStateDiff(Project project, ArDoCoResult arDoCoResult) throws IOException {
@@ -102,7 +75,9 @@ public class ChangedStatesTest {
                 Files.copy(currentPath, previousPath, StandardCopyOption.REPLACE_EXISTING);
             }
         } else {
-            assert Files.exists(previousPath);
+            if (!Files.exists(previousPath)) {
+                throw new IllegalStateException("There is nothing to compare with - please set the overwrite option in ChangedStatesTest!");
+            }
         }
 
         return TextStateFile.writeDiff(previousPath, currentPath, diffPath, arDoCoResult);
@@ -126,12 +101,19 @@ public class ChangedStatesTest {
             var currentPath = evalDir.resolve("recommendationState_" + metaModel.name() + "_" + name + ".txt");
             var previousPath = evalDir.resolve("recommendationState_" + metaModel.name() + "_" + name + "_previous" + ".txt");
             var diffPath = evalDir.resolve("recommendationState_" + metaModel.name() + "_" + name + "_diff" + ".txt");
-            try {
-                Files.copy(currentPath, previousPath, StandardCopyOption.REPLACE_EXISTING);
-            } catch (IOException e) {
-                RecommendationStateFile.write(currentPath, arDoCoResult, metaModel);
-                Files.copy(currentPath, previousPath, StandardCopyOption.REPLACE_EXISTING);
+            if (OVERWRITE_PREVIOUS) {
+                try {
+                    Files.copy(currentPath, previousPath, StandardCopyOption.REPLACE_EXISTING);
+                } catch (IOException e) {
+                    RecommendationStateFile.write(currentPath, arDoCoResult, metaModel);
+                    Files.copy(currentPath, previousPath, StandardCopyOption.REPLACE_EXISTING);
+                }
+            } else {
+                if (!Files.exists(previousPath)) {
+                    throw new IllegalStateException("There is nothing to compare with - please set the overwrite option in ChangedStatesTest!");
+                }
             }
+
             var tempResult = RecommendationStateFile.writeDiff(previousPath, currentPath, diffPath, arDoCoResult, metaModel);
             if (!tempResult)
                 result = tempResult;
@@ -139,7 +121,50 @@ public class ChangedStatesTest {
         return result;
     }
 
-    private static void writeDetailedOutput(Project project, ArDoCoResult arDoCoResult) throws IOException {
+    private static boolean writeConnectionStateDiff(Project project, ArDoCoResult arDoCoResult) throws IOException {
+        String name = project.name().toLowerCase();
+
+        boolean result = true;
+
+        for (var modelId : arDoCoResult.getModelIds()) {
+
+            var metaModel = arDoCoResult.getModelState(modelId).getMetamodel();
+            var evalDir = Path.of(OUTPUT).resolve(name).resolve("connectionStates").resolve(metaModel.name());
+            try {
+                Files.createDirectories(evalDir);
+            } catch (IOException e) {
+                logger.warn("Could not create directories.", e);
+            }
+
+            var currentPath = evalDir.resolve("connectionState_" + metaModel.name() + "_" + name + ".txt");
+            var previousPath = evalDir.resolve("connectionState_" + metaModel.name() + "_" + name + "_previous" + ".txt");
+            var diffPath = evalDir.resolve("connectionState_" + metaModel.name() + "_" + name + "_diff" + ".txt");
+            if (OVERWRITE_PREVIOUS) {
+                try {
+                    Files.copy(currentPath, previousPath, StandardCopyOption.REPLACE_EXISTING);
+                } catch (IOException e) {
+                    ConnectionStateFile.write(currentPath, arDoCoResult, modelId);
+                    Files.copy(currentPath, previousPath, StandardCopyOption.REPLACE_EXISTING);
+                }
+            } else {
+                if (!Files.exists(previousPath)) {
+                    throw new IllegalStateException("There is nothing to compare with - please set the overwrite option in ChangedStatesTest!");
+                }
+            }
+            var tempResult = ConnectionStateFile.writeDiff(previousPath, currentPath, diffPath, arDoCoResult, modelId);
+            if (!tempResult)
+                result = tempResult;
+        }
+        return result;
+    }
+
+    @DisplayName("Text State Diff")
+    @ParameterizedTest(name = "Generating Diff of {0}")
+    @EnumSource(value = Project.class)
+    @Order(2)
+    void textStateDiffTest(Project project) throws IOException {
+        Assumptions.assumeTrue(DATA_MAP.containsKey(project));
+
         String name = project.name().toLowerCase() + "/";
         var path = OUTPUT_PATH.resolve(name);
         try {
@@ -148,8 +173,43 @@ public class ChangedStatesTest {
             logger.warn("Could not create directories.", e);
         }
 
-        Assertions.assertTrue(writeTextStateDiff(project, arDoCoResult));
-        Assertions.assertTrue(writeRecommendationStateDiff(project, arDoCoResult));
+        Assertions.assertTrue(writeTextStateDiff(project, DATA_MAP.get(project)));
+    }
+
+    @DisplayName("Recommendation State Diff")
+    @ParameterizedTest(name = "Generating Diff of {0}")
+    @EnumSource(value = Project.class)
+    @Order(2)
+    void recommendationStateDiffTest(Project project) throws IOException {
+        Assumptions.assumeTrue(DATA_MAP.containsKey(project));
+
+        String name = project.name().toLowerCase() + "/";
+        var path = OUTPUT_PATH.resolve(name);
+        try {
+            Files.createDirectories(path);
+        } catch (IOException e) {
+            logger.warn("Could not create directories.", e);
+        }
+
+        Assertions.assertTrue(writeRecommendationStateDiff(project, DATA_MAP.get(project)));
+    }
+
+    @DisplayName("Connection State Diff")
+    @ParameterizedTest(name = "Generating Diff of {0}")
+    @EnumSource(value = Project.class)
+    @Order(2)
+    void connectionStateDiffTest(Project project) throws IOException {
+        Assumptions.assumeTrue(DATA_MAP.containsKey(project));
+
+        String name = project.name().toLowerCase() + "/";
+        var path = OUTPUT_PATH.resolve(name);
+        try {
+            Files.createDirectories(path);
+        } catch (IOException e) {
+            logger.warn("Could not create directories.", e);
+        }
+
+        Assertions.assertTrue(writeConnectionStateDiff(project, DATA_MAP.get(project)));
     }
 
 }
